// sm_vue_client/src/openapi/system/models/BocSmScriptsDtosCreateScriptDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BocSmScriptsDtosCreateScriptDto = {
    title?: string | null;
    description?: string | null;
    content?: string | null;
    categoryId?: string | null;
};